% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/graph-class.R
\name{MPSEM-generics}
\alias{MPSEM-generics}
\alias{nedge}
\alias{edge}
\alias{edge<-}
\alias{edgenames}
\alias{edgenames<-}
\alias{geometry}
\alias{geometry<-}
\alias{labels<-}
\alias{as.graph}
\alias{locate}
\alias{evolution.model}
\alias{edge.graph}
\alias{edge<-.graph}
\alias{edgenames.graph}
\alias{edgenames<-.graph}
\title{Generic Function (Methods)}
\usage{
nedge(x)

edge(x)

edge(x) <- value

edgenames(x)

edgenames(x) <- value

geometry(x)

geometry(x) <- value

labels(object) <- value

as.graph(x, ...)

locate(x, target, ...)

evolution.model(object, y, ..., x = NULL)

\method{edge}{graph}(x)

\method{edge}{graph}(x) <- value

\method{edgenames}{graph}(x)

\method{edgenames}{graph}(x) <- value
}
\arguments{
\item{x}{A \code{\link{graph-class}} object or an object of any class
that can be converted into a \code{\link{graph-class}} object, or a numeric
vector or numeric matrix of auxiliary trait(s) to help in estimating the
parameters of the trait evolution model. In the second case, the default is
\code{NULL}, which entails that no auxiliary trait is used.}

\item{value}{A two-column \code{\link{data.frame}} giving the coordinates
(e.g., x, y, z) to be applied to the \code{graph-class} object.}

\item{object}{A \code{\link{graph-class}} object or an object of any class
that can be converted into a \code{\link{graph-class}} object.}

\item{...}{Further arguments of the particular conversion method that is
being implemented.}

\item{target}{A numeric or character vector specifying one or more target
species in graph \code{x}.}

\item{y}{A numeric vector or numeric matrix of trait(s) against which to
estimate the parameters of the trait evolution model.}
}
\value{
The return value depends on the implementation of the methods for a
particular S3 class.
}
\description{
A set of non-standard generic functions used by package MPSEM.
}
\section{Functions}{
\itemize{
\item \code{nedge()}: Number of Edges

A method to access the number of edges

\item \code{edge()}: Edge Extraction

A method to extract edges from an object.

\item \code{edge(x) <- value}: Edge Assignment

A method to assign edges to an object.

\item \code{edgenames()}: Edge Names Extraction

A method to extract edge names from an object.

\item \code{edgenames(x) <- value}: Edge Assignment

A method to assign edges to an object.

\item \code{geometry()}: Geometry Extraction

A method to obtain an object's geometry.

\item \code{geometry(x) <- value}: Geometry Assignment

A method to assign a geometry to an object.

\item \code{labels(object) <- value}: Labels Assignment

A method to assign labels to an object.

\item \code{as.graph()}: Graph-class Conversion

A method to convert an object into a graph-class object.

\item \code{locate()}: Locate Targets

A method to locate a set of targets species from a graph-class object.

\item \code{evolution.model()}: Estimate Evolution Model

Estimate the parameters of an evolution model on a PEM2-class object.

\item \code{edge(graph)}: Edge Extraction

Extracts the edges of a graph-class object.

\item \code{edge(graph) <- value}: Edge Assignment

Assigns edges to a graph-class object.

\item \code{edgenames(graph)}: Edge Names Extraction

Extracts the edge names of a graph-class object.

\item \code{edgenames(graph) <- value}: Edge Names Assignment

Assigns edge names to a graph-class object.

}}
